requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.9.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autopep8==2.3.1
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.2.0
cryptography==44.0.2
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dj-database-url==0.5.0
Django==4.2.21
django-allauth==65.8.1
django-crispy-forms==2.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==23.2
django-htmx==1.19.0
django-phonenumber-field==8.0.0
django-redis==5.2.0
django-rename-app==0.1.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
exceptiongroup==1.3.0
# flake8==6.1.0  # 注释掉，解决冲突
# flake8==7.0.0  # 升级到支持更高版本pycodestyle的flake8
frozenlist==1.5.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.4
importlib_resources==6.4.5
inflection==0.5.1
jiter==0.9.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
messagebird==2.2.0
msgpack==1.1.1
multidict==6.1.0
oauthlib==3.2.2
openai==1.88.0
packaging==25.0
paypalrestsdk==1.13.3
phonenumbers==9.0.7
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1  # 保持这个版本
pycparser==2.21
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.1.0
PyJWT==2.8.0
pyOpenSSL==25.0.0
pypng==0.20220715.0
python-alipay-sdk==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
qrcode==7.4.2
redis==4.6.0
referencing==0.35.1
requests==2.32.4
requests-oauthlib==1.3.1
rpds-py==0.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.4
stripe==5.5.0
tomli==2.3.0
tqdm==4.67.1
twilio==9.8.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.0.4
vonage==3.17.2
vonage-jwt==1.1.2
vonage-utils==1.1.3
whitenoise==6.6.0
wrapt==1.17.3
yarl==1.15.2
zipp==3.20.2